requirements.txt
# Основные зависимости для финансового карьерного коуча

# FastAPI и сервер
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# База данных
motor>=3.3.0  # MongoDB async driver

# Валидация данных
pydantic>=2.5.0
pydantic-settings>=2.1.0

# JSON
orjson>=3.9.0

# YandexGPT SDK
yandex-cloud-ml-sdk>=0.15.0
yandexcloud>=0.358.0

# Машинное обучение и поиск
faiss-cpu>=1.12.0  # Быстрый поиск похожих векторов
scikit-learn>=1.7.1
numpy>=1.26.0

# Обработка данных
polars>=1.33.0  # Быстрая обработка данных (как pandas, но быстрее)
pandas>=2.3.2  # Альтернатива polars
pyarrow>=21.0.0  # Для работы с Parquet

# Парсинг и HTTP запросы
requests>=2.31.0
beautifulsoup4>=4.12.0  # Для парсинга HTML (если понадобится)

# Утилиты
tqdm>=4.67.1  # Прогресс-бары
python-dotenv>=1.0.0  # Загрузка .env файлов

# Работа с PDF (для резюме)
pypdf>=5.0.0
reportlab>=4.0.0  # Для экспорта истории в PDF

# Работа с Word документами
python-docx>=1.1.0  # Для экспорта истории в Word

# Логирование
python-json-logger>=2.0.0

# Тестирование (опционально)
pytest>=7.4.0
httpx>=0.25.0  # Для тестирования API

# Форматирование кода (опционально)
ruff>=0.1.0










